Basic working codepoint rendering with metadata text from font files
This squashes a whole bunch of WIP commits from the last month of getting
this working on weekends and evenings. These commits weren't very well
organised, so don't really add any value to keep them in.

This is in a reasonably clean state, but does have a chunk of code for
using harfbuzz commented out in font.cpp as I'm still testing if it's
possible to use harfbuzz for handling glyph substitutions (mainly for
the many emoji that don't have their assigned codepoint).

There's also a number of scripts and test programs that will be removed
once they're deemed no longer necessary. These were from the initial
build where I was investigating pre-rendering all glyphs to an SD card
if there weren't enough resources on the Pico for direct rendering from
font files.