Remove harfbuzz as it's not capable of running on micro memory systems
Harfbuzz's FreeType integration is written to load the entire font into
memory. I wanted Harfbuzz as a drop-in solution for substitution of
emoji sequences, which are stored in the 9MB NotoColorEmoji font
currently. This won't work without significant modification to Harfbuzz,
so I'll need a different solution to achieve ZWJ sequences.

Looks like we'll need to look up the GSUB table on our own to do this 🕶