Use a shared font face between UIFontPen instances to avoid repeated reloading
Loading memory fonts is fast, but it's still a waste of cycles to do it
repeatedly in the same render call. API-wise this isn't great as it adds
a number of places where bugs can creep in, and limits how the UIFontPen
class can be used. These limitations are acceptable with the way it's
currently being used, and the cost of loading fonts.