Fix keys not sending in literal mode
Not sure if I broke this at some point, or if I just never tested it on
the hardware. Sending from literal mode was sending a single Ctrl-Shift-U
and enter, rather than one per character to send.

I've overhauled this so it's more flexible to support other OS sequence
implementations in future.